chore: add .metadata, CHANGELOG.md, and example/advanced_chips_input_example.dart files
The .metadata file is added to track properties of the Flutter project. It is used by the Flutter tool to assess capabilities and perform upgrades. This file should be version controlled and should not be manually edited.

The CHANGELOG.md file is added to keep track of changes made to the project. The initial version 1.0.0 is added.

The example/advanced_chips_input_example.dart file is added as an example of how to use the Flutter Advanced Chips Input package. It demonstrates the usage of the AdvancedChipsInput widget and includes various customization options.